TimeoutExec for vm_mgr, fix bug in feature_extractor where it was commiting from wrong queue, reduce feature extractor threads, reduce active vm to 4 (vbox instability)

+++ vbox-controller/vbox-ctrl.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 
 VBOX_PATH="/cygdrive/c/Program Files/Oracle/VirtualBox"
 
@@ -19,6 +20,11 @@ function manage() {
 
 function stopVm() {
 	local vmName=$1
+
+	if [[ "$(isRunning ${vmName})" -ne "1" ]];then
+	    echo "WARN: vm not running, returning..."
+	    return
+    fi
 
     info "Sending stop signal to vm ${vmName}"
 	manage controlvm ${vmName} poweroff
